+++ udata_piwik/client.py
@@ -8,16 +8,15 @@ from datetime import date
 
 from flask import current_app
 
-log = logging.getLogger(__name__)
+from . import settings
 
-DEFAULT_TRACK_TIMEOUT = 60  # in seconds
-DEFAULT_ANALYZE_TIMEOUT = 60 * 5  # in seconds
+log = logging.getLogger(__name__)
 
 
 def analyze(method, **kwargs):
     """Retrieve JSON stats from Piwik for a given `method` and parameters."""
     base_url = '{0}://{1}/index.php'.format(
-        current_app.config.get('PIWIK_SCHEME', 'http'),
+        current_app.config.get('PIWIK_SCHEME', settings.PIWIK_SCHEME),
         current_app.config['PIWIK_URL'],
     )
     data = {
@@ -35,7 +34,7 @@ def analyze(method, **kwargs):
         kwargs['date'] = dt
     data.update(kwargs)
     timeout = current_app.config.get('PIWIK_ANALYZE_TIMEOUT',
-                                     DEFAULT_ANALYZE_TIMEOUT)
+                                     settings.PIWIK_ANALYZE_TIMEOUT)
     r = requests.get(base_url, params=data, timeout=timeout)
     return r.json()
 
@@ -43,7 +42,7 @@ def analyze(method, **kwargs):
 def track(url, **kwargs):
     """Track a request to a given `url` by issuing a POST against Piwik."""
     base_url = '{0}://{1}/piwik.php'.format(
-        current_app.config.get('PIWIK_SCHEME', 'http'),
+        current_app.config.get('PIWIK_SCHEME', settings.PIWIK_SCHEME),
         current_app.config['PIWIK_URL'],
     )
     data = {
@@ -54,5 +53,5 @@ def track(url, **kwargs):
     }
     data.update(kwargs)
     timeout = current_app.config.get('PIWIK_TRACK_TIMEOUT',
-                                     DEFAULT_TRACK_TIMEOUT)
+                                     settings.PIWIK_TRACK_TIMEOUT)
     requests.post(base_url, data=data, timeout=timeout)